Refactor: Split AdvancedEditor into smaller components
Refactor AdvancedEditor.tsx into smaller, more manageable components to improve maintainability and readability.
